fix compilation
node/detail/node.h  wanted to use 'to_value' function template
in order to convert C string in to std::string, however
that template is defined in node/impl.h which includes
node/detail/node.h so it can not be used without substantial code
reorganization. So we simply make the conversion to std::string.